feat: adds the constructor of a feature to the features querystring if the specific feature does not exist from within polyfill-library
We do not currently have an alias for all the prototype methods that are polyfilled when adding a
polyfill for an entire object such as `Set`, `Map`, `DOMTokenList`. Because of this we currently do
not add those objects to the polyfill.io url when they are present in the features array. This
change adds the ability for the tool to add the constructor/object when it exists and the specific
prototype method polyfill/alias does not exist.